Create feature module and abstract common gradle dependencies
