+++ app/src/main/java/com/kinandcarta/create/proxytoggle/App.kt
@@ -1,0 +1,5 @@
+package com.kinandcarta.create.proxytoggle
+
+import android.app.Application
+
+class App : Application()
