--- app/src/main/java/com/kinandcarta/create/proxytoggle/MainActivity.kt
@@ -1,18 +1,0 @@
-package com.kinandcarta.create.proxytoggle
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.kinandcarta.create.proxytoggle.ui.main.MainFragment
-
-class MainActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.main_activity)
-        if (savedInstanceState == null) {
-            supportFragmentManager.beginTransaction()
-                    .replace(R.id.container, MainFragment.newInstance())
-                    .commitNow()
-        }
-    }
-}
