--- app/src/main/java/com/kinandcarta/create/proxytoggle/ui/main/MainViewModel.kt
@@ -1,7 +1,0 @@
-package com.kinandcarta.create.proxytoggle.ui.main
-
-import androidx.lifecycle.ViewModel
-
-class MainViewModel : ViewModel() {
-    // TODO Implement the ViewModel
-}
