+++ buildScripts/common-dependencies.gradle
@@ -1,0 +1,11 @@
+dependencies {
+    implementation Libs.kotlin_stdlib
+    implementation Libs.core_ktx
+    implementation Libs.appcompat
+    implementation Libs.material
+
+    testImplementation Libs.junit_junit
+
+    androidTestImplementation Libs.androidx_test_ext_junit
+    androidTestImplementation Libs.espresso_core
+}
