+++ buildScripts/common-feature.gradle
@@ -1,0 +1,8 @@
+apply plugin:  'com.android.library'
+apply from: rootProject.file("buildScripts/common.gradle")
+
+dependencies {
+    implementation Libs.constraintlayout
+    implementation Libs.lifecycle_livedata_ktx
+    implementation Libs.lifecycle_viewmodel_ktx
+}
