+++ feature-proxy-manager/build.gradle
@@ -1,0 +1,1 @@
+apply from: rootProject.file("buildScripts/common-feature.gradle")
