+++ feature-proxy-manager/src/main/java/com/kinandcarta/create/proxytoggle/manager/ProxyManagerViewModel.kt
@@ -1,0 +1,7 @@
+package com.kinandcarta.create.proxytoggle.manager
+
+import androidx.lifecycle.ViewModel
+
+class ProxyManagerViewModel : ViewModel() {
+    // TODO Implement the ViewModel
+}
