+++ settings.gradle
@@ -1,2 +1,3 @@
 rootProject.name = "ProxyToggle"
 include ':app'
+include ':feature-proxy-manager'
